+++ lib/src/global_methods.dart
@@ -20,34 +20,7 @@ bool get inAppUpdateSupport => Platform.isAndroid && kReleaseMode;
 
 bool get githubApkReleaseSupport => Platform.isAndroid && kReleaseMode;
 
-void ktShowToast(BuildContext context, String msg,
-    {void Function()? trigger,
-    String actionLabel = 'Undo',
-    Color bgColor = const Color.fromRGBO(0, 20, 41, 0.494)}) {
-  ScaffoldMessenger.of(context).hideCurrentSnackBar();
-  ScaffoldMessenger.of(context).showSnackBar(
-    SnackBar(
-      content: Text(msg,
-          style: Theme.of(context)
-              .textTheme
-              .labelMedium
-              ?.copyWith(color: Colors.white),
-          maxLines: 1,
-          overflow: TextOverflow.ellipsis),
-      backgroundColor: bgColor,
-      behavior: SnackBarBehavior.floating,
-      showCloseIcon: true,
-      action: trigger != null
-          ? SnackBarAction(
-              label: actionLabel,
-              onPressed: trigger,
-            )
-          : null,
-    ),
-  );
-}
-
-Future<bool> ktAuthenticateUser() async {
+Future<bool> ktLocallyAuthenticateUser() async {
   try {
     final auth = LocalAuthentication();
     final isAuthAvailable = await auth.canCheckBiometrics;
@@ -74,34 +47,33 @@ Future<bool> ktRequestPermission(Permission permission) async {
   return false;
 }
 
-Future<({bool isUpdateAvailable, PackageInfo packageInfo})> ktCheckForUpdate(
-    BuildContext context, String repo, String owner, String url) async {
-  final packageInfo = await PackageInfo.fromPlatform();
+Future<
+    ({
+      bool isUpdateAvailable,
+      PackageInfo packageInfo,
+      String downloadUrl,
+      String tag
+    })> ktCheckForUpdate(String repo, String owner, String url) async {
+  final (:packageInfo, :tag) = await ktGetPackageInfo();
+  var isUpdateAvailable = false;
+  var downloadUrl = '';
   if (githubApkReleaseSupport) {
-    final version = packageInfo.version;
-    final buildNumber = packageInfo.buildNumber;
-    final tag = 'v${version.substring(0, 4)}${int.parse(buildNumber) % 1000}';
     final httpClient = HttpClient();
     try {
       final request = await httpClient.postUrl(Uri.parse(url));
       request.headers.set('Content-Type', 'application/x-www-form-urlencoded');
       request.write(Uri(queryParameters: {
-        'repo': 'kaushal',
-        'owner': 'king-technologies',
+        'repo': repo,
+        'owner': owner,
       }).query);
       final response = await request.close();
       if (response.statusCode == 200) {
         final responseBody = await response.transform(utf8.decoder).join();
         final data = jsonDecode(responseBody) as Map<String, dynamic>;
         final networkTag = data['tag_name'] as String;
-        if (tag != networkTag && context.mounted) {
-          ktShowToast(context, 'New version available', trigger: () {
-            launchUrl(
-              Uri.parse(data['downloadUrl'] as String),
-              mode: LaunchMode.externalApplication,
-            );
-          }, actionLabel: 'Download');
-          return (isUpdateAvailable: true, packageInfo: packageInfo);
+        if (tag != networkTag) {
+          isUpdateAvailable = true;
+          downloadUrl = data['downloadUrl'].toString();
         }
       } else {
         debugPrint('Error: ${response.reasonPhrase}');
@@ -111,23 +83,24 @@ Future<({bool isUpdateAvailable, PackageInfo packageInfo})> ktCheckForUpdate(
     } finally {
       httpClient.close();
     }
-    return (isUpdateAvailable: false, packageInfo: packageInfo);
+  } else if (inAppUpdateSupport) {
+    try {
+      final info = await InAppUpdate.checkForUpdate();
+      isUpdateAvailable =
+          info.updateAvailability == UpdateAvailability.updateAvailable;
+    } on Exception catch (e) {
+      debugPrint('Exception: $e');
+    }
   }
-  if (!inAppUpdateSupport) {
-    return (isUpdateAvailable: false, packageInfo: packageInfo);
-  }
-  try {
-    final info = await InAppUpdate.checkForUpdate();
-    final isUpdateAvailable =
-        info.updateAvailability == UpdateAvailability.updateAvailable;
-    return (isUpdateAvailable: isUpdateAvailable, packageInfo: packageInfo);
-  } on Exception catch (e) {
-    debugPrint('Exception: $e');
-  }
-  return (isUpdateAvailable: false, packageInfo: packageInfo);
+  return (
+    tag: tag,
+    isUpdateAvailable: isUpdateAvailable,
+    packageInfo: packageInfo,
+    downloadUrl: downloadUrl
+  );
 }
 
-Future<void> flexibleUpdate() async {
+Future<void> ktFlexibleUpdate() async {
   try {
     final info = await InAppUpdate.checkForUpdate();
     final isUpdateAvailable =
@@ -141,7 +114,14 @@ Future<void> flexibleUpdate() async {
   }
 }
 
-Future<void> immediateUpdate() async {
+Future<({String tag, PackageInfo packageInfo})> ktGetPackageInfo() async {
+  final packageInfo = await PackageInfo.fromPlatform();
+  final tag =
+      'v${packageInfo.version.substring(0, 4)}${int.parse(packageInfo.buildNumber) % 1000}';
+  return (tag: tag, packageInfo: packageInfo);
+}
+
+Future<void> ktImmediateUpdate() async {
   try {
     final info = await InAppUpdate.checkForUpdate();
     final isUpdateAvailable =